refactor: user mapper를 공통 API lib으로 분리
