+++ src/app/api/_lib/user-mapper.ts
@@ -1,0 +1,18 @@
+import type { UserResponse } from '@/contracts/user';
+import type { Database } from '@/lib/supabase/database';
+
+type UsersRow = Database['public']['Tables']['users']['Row'];
+
+export function mapUserRow(row: UsersRow): UserResponse {
+  return {
+    id: row.id,
+    email: row.email,
+    name: row.name,
+    phone: row.phone ?? undefined,
+    profileImage: row.profile_image ?? undefined,
+    role: row.role,
+    status: row.status,
+    createdAt: row.created_at,
+    updatedAt: row.updated_at,
+  };
+}
