+++ src/app/api/users/_lib/mapper.ts
@@ -1,18 +1,1 @@
-import type { UserResponse } from '@/contracts/user';
-import type { Database } from '@/lib/supabase/database';
-
-type UsersRow = Database['public']['Tables']['users']['Row'];
-
-export function mapUserRow(row: UsersRow): UserResponse {
-  return {
-    id: row.id,
-    email: row.email,
-    name: row.name,
-    phone: row.phone ?? undefined,
-    profileImage: row.profile_image ?? undefined,
-    role: row.role,
-    status: row.status,
-    createdAt: row.created_at,
-    updatedAt: row.updated_at,
-  };
-}
+export { mapUserRow } from '@/app/api/_lib/user-mapper';
